pyproject.toml
[project]
name = "close-mongo-ops-manager"
version = "0.5.2"
description = "Monitor and kill MongoDB operations"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "Vicenç Juan Tomàs Monserrat", email = "vicenc.monserrat@close.com" }]
dependencies = [
    "pymongo==4.15.3",
    "pytest-asyncio>=1.2.0",
    "textual>=6.0.0,<7.0.0",
]
keywords = [
  "mongo", "operations", "manager"
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Operating System :: POSIX",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Database :: Front-Ends"
]

[project.urls]
Documentation = "https://github.com/closeio/close-mongo-ops-manager#readme"
Issues = "https://github.com/closeio/close-mongo-ops-manager/issues"
Source = "https://github.com/closeio/close-mongo-ops-manager"

[project.scripts]
close-mongo-ops-manager = "close_mongo_ops_manager.app:main"

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"
